Security-test-suite/requirements.txt
pytest>=8.2
pytest-html>=4.1
pytest-xdist>=3.5
httpx>=0.28
playwright>=1.49
python-dotenv>=1.0
PyJWT>=2.9
pydantic>=2.10
pydantic-settings>=2.7
